Milestone_1/summarization_app/requirements.txt
streamlit==1.9.0
groq==1.0.0
python-dotenv==0.20.0
PyPDF2==1.26.0
reportlab==3.6.2
Pillow==9.0.0
